requirements-dev.txt
pytest
requests
pillow